Decompile `lib` func_us_801BEFEC
