+++ config/splat.us.stlib.yaml
@@ -51,7 +51,8 @@ segments:
       - [0x1A68, .data, e_candle_table]
       - [0x1A90, data]
       - [0x1AD4, .data, e_lesser_demon]
-      - [0x1C48, data]
+      - [0x1C48, .data, e_chair]
+      - [0x1C5C, data]
       - [0x1C7C, .data, st_update]
       - [0x1CA8, .data, collision]
       - [0x2068, .data, e_red_door]
@@ -111,7 +112,6 @@ segments:
       - [0x3B53C, c, unk_3B53C]
       - [0x3B8DC, c, e_lesser_demon]
       - [0x3EDD8, c, e_chair]
-      - [0x3EFEC, c, unk_3EFEC]
       - [0x3F130, c, st_update]
       - [0x3F564, c, collision]
       - [0x40D20, c, create_entity]
