+++ config/splat.us.stno1.yaml
@@ -48,7 +48,8 @@ segments:
       - [0xCFC, .data, st_debug]
       - [0xD7C, .data, unk_35E20]
       - [0xF48, .data, e_mouse]
-      - [0xF8C, data]
+      - [0xF8C, .data, e_chair]
+      - [0xFA0, data]
       - [0x1508, .data, unk_3B5E0]
       - [0x1540, data, unk_3CB34]
       - [0x1570, .data, unk_3D8AC]
@@ -108,7 +109,7 @@ segments:
       - [0x37D24, c, unused]
       - [0x37D34, c, e_mouse]
       - [0x38008, c, e_chair]
-      - [0x381E8, c, unk_381E8]
+      - [0x3832C, c, unk_381E8]
       - [0x3B5E0, c, unk_3B5E0]
       - [0x3C598, c, unk_3CB34]
       - [0x3D8AC, c, unk_3D8AC]
