--- src/st/lib/unk_3EFEC.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "common.h"
-
-INCLUDE_ASM("st/lib/nonmatchings/unk_3EFEC", func_us_801BEFEC);
